requirements.txt
bcrypt==3.2.2
beautifulsoup4==4.11.1
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
click==8.1.3
Flask==2.2.2
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
idna==3.3
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
lxml==4.9.1
MarkupSafe==2.1.1
numpy==1.23.1
pandas==1.4.3
pycparser==2.21
PyJWT==2.4.0
PyMySQL==1.0.2
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.2
requests==2.28.1
six==1.16.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.40
urllib3==1.26.11
Werkzeug==2.2.2
xmltodict==0.13.0
zipp==3.8.1
